requirements.txt
Pillow
sleeper-api-wrapper
requests
